cafeteria-predictor-new/requirements.txt
streamlit
pandas
prophet
azure-storage-blob
openai
python-dotenv